Change progress' bar formula
I figured out that my working formula was a work-around so I decided to make it more scalable and applicable to more applications with different ammount of steps. In this particular project there are 4 steps in total but only 3 remaining since step 1 is already marked as active. So our step cant go up or down by 25%.By substracting 1 from each side I accomplished to change the proportion to the maximum of 3. For example If we are on step 1 and click the next button the progress bar will move by 1/3 of its total, 33%.